Authentication + Authorization

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import {getUserSessionServer} from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma';
 import {NextResponse, NextRequest} from 'next/server';
 import * as yup from 'yup';
@@ -29,10 +30,14 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) {
+    const user = await getUserSessionServer();
+    if (!user) {
+        return NextResponse.json({message: 'Unauthorized'}, {status: 401});
+    }
     try {
 
         const {description, complete} = await postSchema.validate(await request.json());
-        const todo = await prisma.todo.create({data: {description, complete}});
+        const todo = await prisma.todo.create({data: {description, complete, userId: user.id!}});
 
         return NextResponse.json(todo)
 
@@ -44,8 +49,12 @@ export async function POST(request: Request) {
 
 
 export async function DELETE(request: Request) {
+    const user = await getUserSessionServer();
+    if (!user) {
+        return NextResponse.json({message: 'Unauthorized'}, {status: 401});
+    }
     try {
-        await prisma.todo.deleteMany({where: {complete: true}});
+        await prisma.todo.deleteMany({where: {complete: true, userId: user.id!}});
         return NextResponse.json({message: 'Deleted all completed todos'});
     } catch (error) {
         return NextResponse.json(error, {status: 400});
